requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-binance>=1.0.19
ta>=0.11.0
python-dotenv>=1.0.0
plotly>=5.24.0
dash>=2.14.0
